% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{Fit a B/P-spline to data.}
\usage{
bspline(data, formula, degree, smoothing, alpha, knotspacing)
}
\arguments{
\item{degree}{[\code{integer(1)}]\cr
Degree of spline.}

\item{smoothing}{[\code{integer(1)}]\cr
Value of: 0 = none, 1 = identity, 2 = pspline.
Default is 1.}

\item{knotspacing}{[\code{integer(1)}]\cr
Value of: 0 = as points are sampled, 1 = equidistant.
Default is 0.}

\item{xmat}{[\code{matrix(n,p)}]\cr
Matrix of inputs}

\item{y}{[\code{numeric(n)}]\cr
Y vector of outcomes.}
}
\description{
blaaaaaaaaaaaaaaa
}
